Linux & BSD support (#6)
* Implemented From bool for Mode

* Simplified macOS implementation

* Simplified Windows implementation

* Linux & BSD support

The program will try XDG Desktop Portal dbus API (compatible with Flatpak
sandboxing) first to determine current color scheme. 

If that does not work, it will fall back to checking if the theme name contains the string "dark". Checking the theme name is
supported on:
  - Cinnamon
  - GNOME
  - KDE
  - MATE
  - Unity

* Updated README.md for Linux & BSD support

* Reimplemented KDE to use qGray

Added `rgb` method to Mode to implement qGray.

* Reimplemented `detect_kde`

 It now checks `BackgroundNormal` under `Colors:Window`.

* Code cleanup and fixed a typo

* Fix spacing in documentation.

Co-authored-by: Eduardo Flores <edfloreshz@gmail.com>
Co-authored-by: Corey Farwell <coreyf@rwell.org>